refactor: promises utils
